LayoutTests/imported/w3c/web-platform-tests/css/css-shapes/basic-shape-interpolation-expected.txt


FAIL Test circle with negative easing on shape-outside assert_equals: The radius of circle is clamped to zero at 61% expected "circle(0px at 50% 50%)" but got "circle(-0.07151274383068085px at 50% 50%)"
FAIL Test ellipse with negative easing on shape-outside assert_equals: The radius of ellipse is clamped to zero at 61% expected "ellipse(0px 0px at 50% 50%)" but got "ellipse(-0.07151274383068085px -0.07151274383068085px at 50% 50%)"
FAIL Test inset with negative easing on shape-outside assert_equals: The radius of inset is clamped to zero at 61% expected "inset(10%)" but got "inset(10% round -0.07151274383068085px)"
